backend/requirements.txt
# Django
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5

# Database
psycopg2-binary>=2.9
dj-database-url>=2.1

# AWS S3
boto3>=1.34
django-storages>=1.14

# Firebase (for FCM)
firebase-admin>=6.4

# Environment
python-dotenv>=1.0
python-decouple>=3.8

# Utilities
Pillow>=10.2
gunicorn>=21.2

# Development
pytest>=8.0
pytest-django>=4.7
